+++ src/_packages/utils.index.ts
@@ -12,7 +12,7 @@ import { countCharacters } from '../utils/expectation-counters/countCharacters';
 import { countLines } from '../utils/expectation-counters/countLines';
 import { countPages } from '../utils/expectation-counters/countPages';
 import { countParagraphs } from '../utils/expectation-counters/countParagraphs';
-import { countSentences } from '../utils/expectation-counters/countSentences';
+import { countSentences, splitIntoSentences } from '../utils/expectation-counters/countSentences';
 import { countWords } from '../utils/expectation-counters/countWords';
 import { isValidJsonString } from '../utils/isValidJsonString';
 import { extractAllBlocksFromMarkdown } from '../utils/markdown/extractAllBlocksFromMarkdown';
@@ -20,6 +20,23 @@ import { extractAllListItemsFromMarkdown } from '../utils/markdown/extractAllLis
 import { extractOneBlockFromMarkdown } from '../utils/markdown/extractOneBlockFromMarkdown';
 import { removeContentComments } from '../utils/markdown/removeContentComments';
 import { removeMarkdownFormatting } from '../utils/markdown/removeMarkdownFormatting';
+import { capitalize } from '../utils/normalization/capitalize';
+import { decapitalize } from '../utils/normalization/decapitalize';
+import { DIACRITIC_VARIANTS_LETTERS } from '../utils/normalization/DIACRITIC_VARIANTS_LETTERS';
+import { IKeywords, string_keyword } from '../utils/normalization/IKeywords';
+import { isValidKeyword } from '../utils/normalization/isValidKeyword';
+import { nameToUriPart } from '../utils/normalization/nameToUriPart';
+import { nameToUriParts } from '../utils/normalization/nameToUriParts';
+import { normalizeToKebabCase } from '../utils/normalization/normalize-to-kebab-case';
+import { normalizeTo_camelCase } from '../utils/normalization/normalizeTo_camelCase';
+import { normalizeTo_PascalCase } from '../utils/normalization/normalizeTo_PascalCase';
+import { normalizeTo_SCREAMING_CASE } from '../utils/normalization/normalizeTo_SCREAMING_CASE';
+import { normalizeTo_snake_case } from '../utils/normalization/normalizeTo_snake_case';
+import { normalizeWhitespaces } from '../utils/normalization/normalizeWhitespaces';
+import { parseKeywords } from '../utils/normalization/parseKeywords';
+import { parseKeywordsFromString } from '../utils/normalization/parseKeywordsFromString';
+import { removeDiacritics } from '../utils/normalization/removeDiacritics';
+import { searchKeywords } from '../utils/normalization/searchKeywords';
 import { extractBlock } from '../utils/postprocessing/extractBlock';
 import { removeEmojis } from '../utils/removeEmojis';
 import { removeQuotes } from '../utils/removeQuotes';
@@ -30,13 +47,6 @@ import { unwrapResult } from '../utils/unwrapResult';
 // TODO: [🌻] For all, decide if theese are internal or external
 export {
     assertsExecutionSuccessful,
-    countCharacters,
-    countLines,
-    countPages,
-    countParagraphs,
-    countSentences,
-    CountUtils,
-    countWords,
     executionReportJsonToString,
     ExecutionReportStringOptions,
     ExecutionReportStringOptionsDefaults,
@@ -57,27 +67,20 @@ export {
     unwrapResult,
 };
 
+export { countCharacters, countLines, countPages, countParagraphs, countSentences, CountUtils, countWords };
+
+export { splitIntoSentences };
+
 // And the normalization (originally n12 library) utilities:
 
-import { capitalize } from '../utils/normalization/capitalize';
-import { decapitalize } from '../utils/normalization/decapitalize';
-import { DIACRITIC_VARIANTS_LETTERS } from '../utils/normalization/DIACRITIC_VARIANTS_LETTERS';
-import { IKeywords, string_keyword } from '../utils/normalization/IKeywords';
-import { isValidKeyword } from '../utils/normalization/isValidKeyword';
-import { nameToUriPart } from '../utils/normalization/nameToUriPart';
-import { nameToUriParts } from '../utils/normalization/nameToUriParts';
-import { normalizeToKebabCase } from '../utils/normalization/normalize-to-kebab-case';
-import { normalizeTo_camelCase } from '../utils/normalization/normalizeTo_camelCase';
-import { normalizeTo_PascalCase } from '../utils/normalization/normalizeTo_PascalCase';
-import { normalizeTo_SCREAMING_CASE } from '../utils/normalization/normalizeTo_SCREAMING_CASE';
-import { normalizeTo_snake_case } from '../utils/normalization/normalizeTo_snake_case';
-import { normalizeWhitespaces } from '../utils/normalization/normalizeWhitespaces';
-import { parseKeywords } from '../utils/normalization/parseKeywords';
-import { parseKeywordsFromString } from '../utils/normalization/parseKeywordsFromString';
-import { removeDiacritics } from '../utils/normalization/removeDiacritics';
-import { searchKeywords } from '../utils/normalization/searchKeywords';
+export const normalizeTo = {
+    camelCase: normalizeTo_camelCase,
+    PascalCase: normalizeTo_PascalCase,
+    'SCREAMING-CASE': normalizeTo_SCREAMING_CASE,
+    snake_case: normalizeTo_snake_case,
+    'kebab-case': normalizeToKebabCase,
+};
 
-// TODO: !!! Try to Export also snake-case ,...
 export {
     capitalize,
     decapitalize,