+++ src/utils/expectation-counters/countWords.ts
@@ -1,19 +1,6 @@
 import type { ExpectationAmount } from '../../types/PromptbookJson/PromptTemplateJson';
 import { removeDiacritics } from '../normalization/removeDiacritics';
 
-
-/**
- * Counts number of words in the text
- */
-export function countWords(text: string): ExpectationAmount {
-  text = text.replace(/[\p{Extended_Pictographic}]/gu, 'a');
-  text = removeDiacritics(text);
-
-  return text.split(/[^a-zа-я0-9]+/i).filter((word) => word.length > 0).length;
-}
-
-
-
 /**
  * Counts number of words in the text
  */