+++ src/types/PromptbookJson/PromptTemplateJson.ts
@@ -32,6 +32,17 @@ export type NaturalTemplateJson = PromptTemplateJsonCommon & {
 };
 
 /**
+ * Expect this amount of each unit in the answer
+ *
+ * For example 5 words, 3 sentences, 2 paragraphs, ...
+ *
+ * Note: Expectations are performed after all postprocessing steps
+ */
+export type Expectations = Partial<
+    Record<Lowercase<ExpectationUnit>, { min?: ExpectationAmount; max?: ExpectationAmount }>
+>;
+
+/**
  * Units of text measurement
  */
 export const EXPECTATION_UNITS = ['CHARACTERS', 'WORDS', 'SENTENCES', 'PARAGRAPHS', 'LINES', 'PAGES'] as const;
@@ -133,9 +144,7 @@ interface PromptTemplateJsonCommon {
      *
      * Note: Expectations are performed after all postprocessing steps
      */
-    readonly expectAmount?: Partial<
-        Record<Lowercase<ExpectationUnit>, { min?: ExpectationAmount; max?: ExpectationAmount }>
-    >;
+    readonly expectations?: Expectations;
 
     /**
      * Expect this format of the answer